Store logs in text files (#7)
* trying with textfiles

* tests for text files

* add test for instance constructor

* add test for constructor

* move file path to helper method

* sleep for less ms

+++ src/main/java/com/simplymadeapps/simple_logger_android/SimpleAmazonLogs.java
@@ -1,11 +1,9 @@
 package com.simplymadeapps.simple_logger_android;
 
 import android.app.Application;
-import android.arch.persistence.room.Room;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
-import android.util.Log;
 
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.mobileconnectors.s3.transferutility.TransferListener;
@@ -16,12 +14,11 @@ import com.amazonaws.regions.Region;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
+import com.snatik.storage.Storage;
 
 import java.io.File;
-import java.io.FileOutputStream;
-import java.util.ArrayList;
-import java.util.Calendar;
-import java.util.GregorianCalendar;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 
 public class SimpleAmazonLogs {
@@ -44,11 +41,14 @@ public class SimpleAmazonLogs {
     protected static long last_clear_old_logs_checked = 0;
     final static long HRS_24_IN_MS = 86400000l;
 
-    protected static RecordedLogDatabase database;
+    protected static Storage storage;
+    protected static String storagePath;
+
+    protected static SimpleDateFormat date_format_log = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss z");
+    protected static SimpleDateFormat date_format_title = new SimpleDateFormat("yyyy-MM-dd");
 
     /**
      * @param application the application object used for storing records
-     * @return the entry point for adding, reading, and uploading logs
      */
     public static void init(Application application) {
         if(instance == null) {
@@ -58,73 +58,63 @@ public class SimpleAmazonLogs {
 
     // Constructor
     protected SimpleAmazonLogs(Application application) {
-        this.database = Room.databaseBuilder(application, RecordedLogDatabase.class, "recorded_log_database").build();
+        this.storage = SimpleAmazonLogsHelper.newStorageInstance(application.getApplicationContext());
         this.context = application.getApplicationContext();
         this.preferences = PreferenceManager.getDefaultSharedPreferences(context);
         this.editor = preferences.edit();
         this.daysToKeepInStorage = preferences.getInt(KEEP_IN_STORAGE_KEY, 7);
+        this.storagePath = storage.getInternalFilesDirectory()+File.separator+"SIO-Logs";
     }
 
-    /**
-     * @param log takes a string and stores it as a log.  The time stamp will be added automatically.
-     */
-    public static void addLog(final String log) {
-        Thread thread = new Thread() {
-            @Override
-            public void run() {
-                if(haveNotCheckedForOldLogsInLast24Hrs()) { // Only check to clear old logs if we haven't in the last 24 hrs
-                    clearOldLogs();  // Because adding a log is called frequently, we will use it to check for week old logs and delete them
-                }
+    public static void addLog(String log) {
+        if(haveNotCheckedForOldLogsInLast24Hrs()) { // Only check to clear old logs if we haven't in the last 24 hrs
+            clearOldLogs();  // Because adding a log is called frequently, we will use it to check for week old logs and delete them
+        }
 
-                RecordedLog recordedLog = new RecordedLog();
-                recordedLog.setLog(log);
-                recordedLog.setDate(System.currentTimeMillis());
-                database.recordedLogDao().insert(recordedLog);
-            }
-        };
-        thread.start();
+        Date date = new Date(System.currentTimeMillis());
+        String textFileName = date_format_title.format(date);
+        String logDateTag = date_format_log.format(date);
+        storage.createDirectory(storagePath, false);
+        createTextFileIfItDoesntExist(textFileName);
+        storage.appendFile(getFilePath(textFileName), logDateTag+" - "+log+"");
     }
 
-    /**
-     * @return returns true if we have already checked for old logs today, false if not
-     */
+    protected static void createTextFileIfItDoesntExist(String fileName) {
+        boolean exists = storage.isFileExist(getFilePath(fileName));
+        if(!exists) {
+            storage.createFile(getFilePath(fileName), "");
+        }
+    }
+
+    protected static String getFilePath(String fileName) {
+        return storagePath+File.separator+fileName;
+    }
+
     protected static boolean haveNotCheckedForOldLogsInLast24Hrs() {
         return last_clear_old_logs_checked < System.currentTimeMillis()-HRS_24_IN_MS;
     }
 
-    /**
-     * @return returns all currently stored logs
-     */
-    public static void getAllLogs(final RecordedLogCallbacks callback) {
-        Thread thread = new Thread() {
-            @Override
-            public void run() {
-                callback.onLogsReady(database.recordedLogDao().getAll());
-            }
-        };
-        thread.start();
-    }
-
-    /**
-     * @param days sets how long we should keep logs stored locally - by default, logs are deleted after 7 days
-     */
     public static void setStorageDuration(int days) {
         editor.putInt(KEEP_IN_STORAGE_KEY, days);
         editor.commit();
         daysToKeepInStorage = days;
     }
 
-    /**
-     * Deletes all stored records
-     */
     public static void deleteAllLogs() {
-        Thread thread = new Thread() {
-            @Override
-            public void run() {
-                database.recordedLogDao().deleteAll();
-            }
-        };
-        thread.start();
+        storage.deleteDirectory(storagePath);
+    }
+
+    public static String getLogsFromDaysAgo(int daysAgo) {
+        long decrement = daysAgo*HRS_24_IN_MS;
+        String current_day_title = date_format_title.format(new Date(System.currentTimeMillis()-decrement));
+        boolean exists = storage.isFileExist(getFilePath(current_day_title));
+        if(exists) {
+            return storage.readTextFile(getFilePath(current_day_title));
+        }
+        else {
+            // There are no logs for the current day
+            return "";
+        }
     }
 
     /**
@@ -143,66 +133,15 @@ public class SimpleAmazonLogs {
     // Clear logs from storage that are older than the threshold
     protected static void clearOldLogs() {
         long decrement = daysToKeepInStorage*HRS_24_IN_MS;
-        // no need for it is own thread as this always gets called from the addLogs thread
-        database.recordedLogDao().deleteLogsOlderThanTime(System.currentTimeMillis()-decrement);
         last_clear_old_logs_checked = System.currentTimeMillis();
-    }
-
-    // Will return a text file with all of the logs.  It will return null if there was an error creating the file
-    protected static File createLogTextFile(List<RecordedLog> logs) {
-        try {
-            File file = File.createTempFile(System.currentTimeMillis()+"", ".txt");
-            FileOutputStream stream = new FileOutputStream(file);
-            for(RecordedLog log : logs) {
-                String line = log.getReadableDate() + " - " + log.getLog() + "\n";
-                stream.write(line.getBytes());
-            }
-            stream.close();
-            return file;
-        }
-        catch(Exception e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    // Get the logs from a specific day
-    // The range will end up being as follows:
-    // 06/25 12:00:00 <= date < 06/26 12:00:00
-    // So it should give all logs perfectly for that date
-    protected static List<RecordedLog> getLogsFromSpecificDay(int daysAgo) {
-        // get the date/time for right now
-        Calendar date = new GregorianCalendar();
-
-        // reset hour, minutes, seconds and millis so it is midnight
-        date.set(Calendar.HOUR_OF_DAY, 0);
-        date.set(Calendar.MINUTE, 0);
-        date.set(Calendar.SECOND, 0);
-        date.set(Calendar.MILLISECOND, 0);
-
-        // This will set the date to daysAgo at midnight
-        date.add(Calendar.DAY_OF_MONTH, (daysAgo*-1)+1);
-        long upperBound = date.getTimeInMillis();
-        Log.d("UPPER BOUND",upperBound+"");
-
-        // This will set the date to daysAgo+1 at midnight
-        date.add(Calendar.DAY_OF_MONTH, -1);
-        long lowerBound = date.getTimeInMillis();
-        Log.d("LOWER BOUND",lowerBound+"");
-
-        return database.recordedLogDao().getLogsWithinTimeRange(lowerBound, upperBound);
-    }
-
-    // Get list of list of logs to be uploaded - one list of logs per day
-    protected static List<List<RecordedLog>> getListOfListOfLogsToUpload() {
-        List<List<RecordedLog>> list_of_lists = new ArrayList<>();
-        for(int i = 0; i < daysToKeepInStorage; i++) {
-            List<RecordedLog> logs_from_day = getLogsFromSpecificDay(i);
-            if(logs_from_day.size() > 0) {
-                list_of_lists.add(logs_from_day);
+        List<File> files = storage.getFiles(getFilePath(""));
+        if(files != null) {
+            for(File file : files) {
+                if(file.lastModified() < System.currentTimeMillis()-decrement) {
+                    file.delete();
+                }
             }
         }
-        return list_of_lists;
     }
 
     protected static boolean verifyAmazonCredentialsHaveBeenAdded() {
@@ -219,84 +158,53 @@ public class SimpleAmazonLogs {
             callback.onFailure(new Exception("You must call setAmazonCredentials() before uploading to Amazon"), 0, 0);
         }
         else {
-            Thread thread = new Thread() {
-                @Override
-                public void run() {
-                    // Create an S3 client
-                    AmazonS3 s3 = new AmazonS3Client(new BasicAWSCredentials(access_token,secret_token));
+            // Create an S3 client
+            AmazonS3 s3 = new AmazonS3Client(new BasicAWSCredentials(access_token,secret_token));
 
-                    // Set the region of your S3 bucket
-                    s3.setRegion(Region.getRegion(region));
+            // Set the region of your S3 bucket
+            s3.setRegion(Region.getRegion(region));
 
-                    // We have to do an upload for each day of logs we have.  This could mean there is between 0 and 7 logs that need to be uploaded
-                    List<List<RecordedLog>> list_of_list_of_logs = getListOfListOfLogsToUpload();
-                    System.out.println(list_of_list_of_logs.size());
+            final List<File> files_to_upload = storage.getFiles(storagePath);
+            final int TOTAL_LOGS_TO_UPLOAD = files_to_upload.size();
+            successful_calls = 0;
+            unsuccessful_calls = 0;
 
-                    final int TOTAL_LOGS_TO_UPLOAD = list_of_list_of_logs.size();
-                    successful_calls = 0;
-                    unsuccessful_calls = 0;
+            for(File file : files_to_upload) {
+                // Upload to amazon
+                TransferUtility transferUtility = new TransferUtility(s3, context);
+                TransferObserver observer = TransferHelper.getTransferObserver(transferUtility, directory, bucket, file.getName(), file);
 
-                    for(List<RecordedLog> list_of_logs : list_of_list_of_logs) {
-                        final File file = createLogTextFile(list_of_logs); // Create a text file from the logs
-                        String filename = list_of_logs.get(0).getTextFileTitle(); // Generate a name for the file
-
-                        // Upload to amazon
-                        TransferUtility transferUtility = new TransferUtility(s3, context);
-                        TransferObserver observer = TransferHelper.getTransferObserver(transferUtility, directory, bucket, filename, file);
-
-                        // Monitor the upload so we can callback when completed
-                        observer.setTransferListener(getTransferListener(TOTAL_LOGS_TO_UPLOAD, list_of_logs, file, callback));
-                    }
-
-                    // There was nothing to upload - let's just return success
-                    if(TOTAL_LOGS_TO_UPLOAD == 0) {
-                        callback.onSuccess(0);
-                    }
-                }
-            };
-            thread.start();
-        }
-    }
-
-    public static void deleteListOfLogs(final List<RecordedLog> list_of_logs) {
-        Thread thread = new Thread() {
-            @Override
-            public void run() {
-                database.recordedLogDao().delete(list_of_logs);
+                // Monitor the upload so we can callback when completed
+                observer.setTransferListener(getTransferListener(TOTAL_LOGS_TO_UPLOAD, file, callback));
             }
-        };
-        thread.start();
+
+            // There was nothing to upload - let's just return success
+            if(TOTAL_LOGS_TO_UPLOAD == 0) {
+                callback.onSuccess(0);
+            }
+        }
     }
 
     protected static int successful_calls = 0;
     protected static int unsuccessful_calls = 0;
 
     // This is our transfer listener for when we upload the file
-    protected static TransferListener getTransferListener(final int TOTAL_LOGS_TO_UPLOAD, final List<RecordedLog> list_of_logs, final File file, final SimpleAmazonLogCallback callback) {
+    protected static TransferListener getTransferListener(final int TOTAL_LOGS_TO_UPLOAD, final File file, final SimpleAmazonLogCallback callback) {
         return new TransferListener() {
             @Override
             public void onStateChanged(int id, TransferState state) {
                 if (state.equals(TransferState.COMPLETED)) {
-                    // We want to delete this temporary file
-                    file.delete();
-
                     // Iterate the number of uploads that have completed
                     successful_calls = successful_calls + 1;
 
-                    // This just gets me a dummy log for today's date
-                    RecordedLog recordedLog = new RecordedLog();
-                    recordedLog.setDate(System.currentTimeMillis());
-
-                    // If the uploaded logs are from a previous day (not from today) we can clear those logs.
-                    if(!list_of_logs.get(0).getTextFileTitle().equalsIgnoreCase(recordedLog.getTextFileTitle())) {
-                        deleteListOfLogs(list_of_logs);
+                    String current_day_title = date_format_title.format(new Date(System.currentTimeMillis()));
+                    if(!current_day_title.equalsIgnoreCase(file.getName())) {
+                        // We can delete any text file that isn't from today
+                        file.delete();
                     }
                 }
 
                 if (state.equals(TransferState.CANCELED) || state.equals(TransferState.FAILED)) {
-                    // We want to delete this temporary file
-                    file.delete();
-
                     // Failed upload
                     unsuccessful_calls = unsuccessful_calls + 1;
                 }
@@ -322,7 +230,6 @@ public class SimpleAmazonLogs {
             @Override
             public void onError(int id, Exception ex) {
                 // Failed upload
-                file.delete();
                 ex.printStackTrace();
                 unsuccessful_calls = unsuccessful_calls + 1;
                 if(successful_calls + unsuccessful_calls == TOTAL_LOGS_TO_UPLOAD) {
